// lib/faust-types.ts
export interface AboutPageFieldsType {
  sectionTitle: string;
  sectionTitleItalic: string;
  nepaliDescription: string;
  imageStack: AboutImageStack[];
}

export interface AboutFieldNode {
  aboutPageFieldsType: AboutPageFieldsType;
}

export interface AboutSectionData {
  aboutFields: {
    nodes: AboutFieldNode[];
  };
}